# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorModel": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorModel",
        "type": "object"
      },
...
      "SessionModel": {
        "properties": {
          "kind": {
            "default": "guest",
            "title": "Kind",
            "type": "string"
          }
        },
        "title": "SessionModel",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Interactive TV series experience: search, episode tracking, comments and AI powered insights over the public TVMaze API.",
    "title": "tv-insight",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/session": {
      "get": {
        "description": "Confirms the session model without exposing the identifier.\n\nThe cookie is httpOnly on purpose; the client only needs to know that it is\nbrowsing as a guest.",
        "operationId": "read_session_api_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unknown resource"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Upstream catalogue failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "No insight provider available"
          }
        },
        "summary": "Describe the current session",
        "tags": [
          "session"
        ]
      }
    },
    "/api/session/reset": {
      "post": {
        "description": "Discard the current guest identity.\n\nDeliberately does not mint the replacement here: the next request that needs a\nviewer will create one, which keeps the identity logic in a single place\n(``dependencies.get_viewer_id``).",
        "operationId": "reset_session_api_session_reset_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unknown resource"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Upstream catalogue failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "No insight provider available"
          }
        },
        "summary": "Start a new guest session",
        "tags": [
          "session"
        ]
      }
    }
  }
}